Added code needed for demo in a branch for now.
The jpylyzer wrapper is a java / sax based wrapper for the CL jpylyzer.
The Jp2StructChecker does the same thing for the struct checking tool.
ContainerUnpcaker unpacks zips and allows the entries to be accessed as files.
The Jhove Characterisation wrapper is a clunky jhove invoker.

These aren't production tools yet.